requirements.txt
flask==3.0.0
bcrypt>=4.0.0
flask-wtf>=1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
markdown==3.5.1
bleach==6.1.0
anthropic>=0.18.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
